[plaster] Initial prototype
This is the initial prototype for 🩹 Plaster. This work is based on a
recenlty worked out proposal:

https://docs.google.com/document/d/1aPf7cMB5Pc_ssKFC_I4-mLHr-7D4lTqsvItSXY-WzQM/

Run this tool with:

```
tools/cr/plaster.py
```

The tool at the moment is in a very early prototyping stage, and there
are many tickets opened for it, to get it into a mature state. The
intent is to manage to deliver this in gradual steps, while acquiring
experience alongside the way.

- Overlay files.

These are python notation files, even though they look pretty similar to
JSON. The advantage of these files is that they will allow very flexible
composition of regular expressions.

The `plasters` array is currently being dedicated to the different
regexes to be applied to a file. The format for plasters is
something like this

```python
{
   "pattern": r"",
   "replacement": r"",
   "flags": [],
}
```

A user can list as many of these entries as necessary for each regex
that one desires to apply to a given source.

- Templates

As a lot of the patches are similar operations (e.g tagging functions
virtual, making classes friend), 🩹 will support a template approach,
where the plaster record is provided as a template that can be used
inside the overlay file. This means we can have ever more robust regexes
for the changes we want to apply, and reuse them across multiple files.

- How does it work

An overlay file has to have the same path of the source file, but under
`rewrite`, but with the `.overlay` extension added to it. This is
similar to what we do to `chromium_src` shadow files.

When running 🩹, all overlay files are visited, the version of the
source in git is retrieved from its repository, and the changes are
applied to that version. Once a file is processed, the changes are saved
to the source in the tree, and a patch file is generated based on it.

Resolves https://github.com/brave/brave-browser/issues/45051